% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildEnrichMatrix.R
\name{BuildEnrichMatrix}
\alias{BuildEnrichMatrix}
\title{BuildEnrichMatrix}
\usage{
BuildEnrichMatrix(genes, type = "GO", db = NULL, species = "Human")
}
\arguments{
\item{type}{}

\item{db}{}
}
\description{
Build enrichment matrix
}
