% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMSigDB.R
\name{BuildMSigDB}
\alias{BuildMSigDB}
\title{BuildMSigDB}
\usage{
BuildMSigDB(type, species)
}
\arguments{
\item{type}{GO, KEGG, HALLMARK, MOTIF, PATHWAYS, BIOCARTA, REACTOME, CELLTYPE.}

\item{species}{Human or Mouse}
}
\description{
Create a reference MSigDB
}
