% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{Enrichment}
\alias{Enrichment}
\title{Enrichment}
\usage{
Enrichment(
  geneset,
  genes = NULL,
  db = NULL,
  enrichment.matrix = NULL,
  seurat_obj = NULL,
  type = "GO",
  species = "Human",
  do.sort = FALSE,
  pval.thresh = 1,
  pval.n = Inf,
  do.plot = F,
  do.print = F
)
}
\arguments{
\item{do.print}{}
}
\description{
Enrichment
}
