% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotEnrichment.R
\name{PlotEnrichment}
\alias{PlotEnrichment}
\title{PlotEnrichment}
\usage{
PlotEnrichment(pvals, do.sort = F, pval.thresh = 1, pval.n = Inf, do.print = F)
}
\arguments{
\item{...}{}
}
\description{
PlotEnrichment
}
