% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scEnrichment.R
\name{scEnrichment}
\alias{scEnrichment}
\title{scEnrichment}
\usage{
scEnrichment(
  markers,
  genes = NULL,
  db = NULL,
  seurat_obj = NULL,
  enrich.method = "clusterProfiler",
  species = "Human",
  type = "GO",
  do.sort = FALSE,
  pval.thresh = 1,
  pval.n = Inf,
  do.plot = F,
  do.print = F
)
}
\arguments{
\item{markers}{Output of `FindMarkers` or `FindAllmarkers` in Seurat package.}

\item{type}{}

\item{output}{}
}
\description{
Perform enrichment analysis on single-cell level
}
